refactor(#9): annotate() 内部実装を core/annotation_runner.py に分離
Issue #9 のリファクタ: api.annotate() の複数モデル実行ロジックを内部モジュール
core/annotation_runner.py に切り出す。

## 変更内容

- core/annotation_runner.py を新規作成 (~400 行)
  - インスタンス管理 (_MODEL_INSTANCE_REGISTRY, get_annotator_instance,
    _create_annotator_instance, PydanticAIWebAPIWrapper など)
  - 実行制御 (_prepare_phash_map, _annotate_model, _execute_model_annotation)
  - 結果格納 (_process_model_results, _handle_error, _record_model_error)
  - 公開関数 run_annotation() を新設 (旧 annotate() のループ部分)
- api.py を 526 行 → 50 行に縮減
  - annotate() は run_annotation() への 1 行委譲
  - list_available_annotators() のみ残置
- core/types.py に PHashAnnotationResults を移動 (循環import回避のため)
  - api.py は core/types.py から re-import (公開シンボルとして維持)
- tests/unit/fast/test_api.py のモックパスを新モジュールに更新
- tests/unit/fast/test_annotation_runner.py を新設 (smoke test 6 件)

## 設計判断

Issue 本文の例は `class AnnotationRunner` だが、状態を持たない callable class
パターンになるため、Python idiom として **モジュール関数** で実装。状態を持つ
拡張が必要になれば関数 → クラスへリスクなく昇格可能。

## 互換性

- 公開 API `image_annotator_lib.api.annotate()` のシグネチャと挙動は完全互換
- 公開シンボル `PHashAnnotationResults` も維持 (api.py から re-export)
- 既存テスト 56/56 PASS (image-annotator-lib fast unit tests)

## 既知の事項

mypy エラー 5 件 (annotator_class の型ヒント欠落、SimplifiedAgentWrapper の
api_keys 引数) は元 api.py からの完全コピー部分にあり、Issue #9 のスコープ外。
